+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -353,6 +353,8 @@ public class SampleSQLBuilder {
 
 	private Map<String, Object> _createContext() throws Exception {
 		return HashMapBuilder.<String, Object>put(
+			"blogDataFactory", new BlogDataFactory()
+		).put(
 			"dataFactory", new DataFactory()
 		).put(
 			"dDLDDMDataFactory", new DDLDDMDataFactory()
